Input과 관련된 State와 함수를 todoInput 컴포넌트에 저장
