+++ components/todo-input.js
@@ -1,11 +1,28 @@
-import React from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 
 export default function TodoInput(props) {
-  const { date, todo, onChangeTodo, onChangeDate, addItem } = props;
+  const { addItem, initForm } = props;
+
+  const [date, setDate] = useState("");
+  const [todo, setTodo] = useState("");
+
+  const onChangeTodo = (e) => {
+    setTodo(e.target.value);
+  };
+  const onChangeDate = (e) => {
+    //입력을 8자리 넘어가면 받지 않도록 설정
+    if (String(e.target.value).length <= 8) setDate(e.target.value);
+  };
+
   // value 값이 처음에 없어서 undefined가 되면 오류가 발생할 수 있다고하므로, || ""을 추가해준다.
   return (
-    <Wrapper onSubmit={addItem}>
+    <Wrapper
+      onSubmit={() => {
+        addItem(event, date, todo);
+        initForm(setDate, setTodo);
+      }}
+    >
       <InputWrapper>
         <InputLabel>시간</InputLabel>
         <DateInput
